pyproject.toml
[tool.poetry]
name = "viaduct"
version = "1.0"
description = "Rate-limited API Challenge"
authors = ["m4s0 <masotti79@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
catboost = "0.24"
pytest = "^6.2.2"
fastapi = "^0.63.0"
uvicorn = "^0.13.4"
requests = "^2.25.1"
mysql-connector-python = "8.0.22"

[tool.poetry.dev-dependencies]
mypy = "0.761"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
